requirements.txt
# local package
-e .

# external requirements
google-search-results
pandas
snorkel
duckdb
delta
delta-spark
deltalake
pyspark
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
